Trying half precision - it doesn't work on cpu hopefully it does on GPU

+++ mpnn_functions/readout/graph_level_output.py
@@ -34,7 +34,7 @@ class GraphLevelOutput(nn.Module):
         if mask is None:
             gated_activations = self.attn_act(self.i(input_set)).mul(self.j(input_set))
         else:
-            att_mask = mask.float()
+            att_mask = mask.half()
             att_mask = (1 - att_mask) * _BIG_NEGATIVE
             gated_activations = self.attn_act(self.i(input_set) + att_mask).mul(self.j(input_set)).mul(mask)
         return gated_activations.sum(dim=1)
